test(cmake): add test for installable with `C++17`
