+++ .github/workflows/build.yml
@@ -39,8 +39,8 @@ concurrency:
 jobs:
   build-sdist:
     name: Build sdist
-    runs-on: ubuntu-latest
     if: github.repository == 'metaopt/optree' && (github.event_name != 'push' || startsWith(github.ref, 'refs/tags/'))
+    runs-on: ubuntu-latest
     timeout-minutes: 15
     steps:
       - name: Checkout
@@ -81,9 +81,9 @@ jobs:
 
   build-wheels:
     name: Build wheels for Python ${{ matrix.python-version }} on ${{ matrix.os }} (${{ matrix.archs }})
+    if: github.repository == 'metaopt/optree' && (github.event_name != 'push' || startsWith(github.ref, 'refs/tags/'))
     runs-on: ${{ matrix.os }}
     needs: [build-sdist]
-    if: github.repository == 'metaopt/optree' && (github.event_name != 'push' || startsWith(github.ref, 'refs/tags/'))
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
@@ -200,9 +200,9 @@ jobs:
 
   list-artifacts:
     name: List artifacts
+    if: github.repository == 'metaopt/optree' && (github.event_name != 'push' || startsWith(github.ref, 'refs/tags/'))
     runs-on: ubuntu-latest
     needs: [build-sdist, build-wheels]
-    if: github.repository == 'metaopt/optree' && (github.event_name != 'push' || startsWith(github.ref, 'refs/tags/'))
     timeout-minutes: 15
     steps:
       - name: Download built sdist
@@ -243,8 +243,8 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Set up Python
-        uses: actions/setup-python@v5
         if: startsWith(github.ref, 'refs/tags/')
+        uses: actions/setup-python@v5
         with:
           python-version: "3.7 - 3.13"
           update-environment: true
