+++ src/treespec/richcomparison.cpp
@@ -19,11 +19,11 @@ limitations under the License.
 #include <unordered_map>  // std::unordered_map
 #include <vector>         // std::vector
 
-#include "include/critical_section.h"
 #include "include/exceptions.h"
 #include "include/pytypes.h"
 #include "include/registry.h"
 #include "include/stdutils.h"
+#include "include/synchronization.h"
 #include "include/treespec.h"
 
 namespace optree {
